Split out Publication metadata
