+++ src/lib/components/Publication.svelte
@@ -1,5 +1,6 @@
 <script lang="ts" strictEvents>
   import CodeBlock from "$lib/components/CodeBlock.svelte";
+  import Metadata from "$lib/components/Publication/Metadata.svelte";
 
   export let abstract: string;
   export let authors: Array<string>;
@@ -14,58 +15,12 @@
   let showBib = false;
 
   $: descriptionWidth = previewImage !== undefined ? "50%" : "85%";
-
-  function monthName(month: number): string {
-    switch (month) {
-      case 1:
-        return "January";
-      case 2:
-        return "February";
-      case 3:
-        return "March";
-      case 4:
-        return "April";
-      case 5:
-        return "May";
-      case 6:
-        return "June";
-      case 7:
-        return "July";
-      case 8:
-        return "August";
-      case 9:
-        return "September";
-      case 10:
-        return "October";
-      case 11:
-        return "November";
-      case 12:
-        return "December";
-    }
-    return month.toString();
-  }
-
-  function formatDate(date: Date): string {
-    return (
-      date.getDay() +
-      1 +
-      " " +
-      monthName(date.getMonth() + 1) +
-      " " +
-      date.getFullYear()
-    );
-  }
 </script>
 
 <div class="row">
   <div class="container">
     <div class="metadata">
-      <div>{formatDate(date)}</div>
-      <div>
-        {#each tags as tag, i (i)}
-          <div class="tag">{tag}</div>
-        {/each}
-      </div>
+      <Metadata {date} {tags} />
     </div>
     <div style:flex="0 0 {descriptionWidth}" style:width={descriptionWidth}>
       <h2>{title}</h2>
@@ -181,7 +136,6 @@
 
   .metadata {
     flex: 0 0 15%;
-    font-size: 0.8rem;
     padding-top: 0.5rem;
   }
 
@@ -206,17 +160,5 @@
 
   .source-buttons {
     display: flex;
-  }
-
-  .tag {
-    background-color: var(--primary-bg-color);
-    border: 1px solid var(--primary-color);
-    border-radius: 3px;
-    margin-top: 0.5rem;
-    padding: 0.1rem 0.3rem;
-    width: fit-content;
-    transition:
-      background-color theme.$transition-duration ease,
-      border-color theme.$transition-duration ease;
   }
 </style>
